added write_index module to antiweb_lib package
